fix: apply @julianken-bot review - use single client component, gate on mounted
- BLOCKER: drop next/dynamic wrapper; MermaidDiagram.tsx is now a single
  'use client' component; MermaidDiagram.client.tsx deleted
- IMPORTANT: gate render on mounted (useSyncExternalStore pattern) so
  dark-mode readers never see a light-theme flash on first paint
- initialize-race: module-level lastInitializedTheme guard; mermaid.initialize
  only called when theme value changes, not on every source re-render
